AG-8890 - Label and marker util functions
