+++ packages/ag-charts-enterprise/src/rangeArea/typings.ts
@@ -19,12 +19,11 @@ export interface AgRangeAreaSeriesMarkerFormatterParams<DatumType> {
     readonly size: number;
     readonly fill?: CssColor;
     readonly stroke?: CssColor;
-    readonly strokeWidth: PixelSize;
+    readonly strokeWidth?: PixelSize;
     readonly highlighted: boolean;
     readonly xKey: string;
     readonly yLowKey: string;
     readonly yHighKey: string;
-    readonly labelKey?: string;
     readonly seriesId: string;
     readonly itemId: string;
 }
